refactor(gh-pages): cleanup page.less, split variables.less
