+++ pages/src/common/breakpoints.less
@@ -1,0 +1,27 @@
+/* Breakpoints definition */
+@xs-min: 1px;
+@xs-max: 599px;
+@sm-min: @xs-max + 1;
+@sm-max: 991px;
+@md-min: @sm-max + 1;
+@md-max: 1199px;
+@lg-min: @md-max + 1;
+@lg-max: 1599px;
+@xl-min: @lg-max + 1;
+@xl-max: 999999px;
+
+/* Media queries */
+@xs: ~'(min-width: @{xs-min}) and (max-width: @{xs-max})';
+@xs-sm: ~'(min-width: @{xs-min}) and (max-width: @{sm-max})';
+@xs-md: ~'(min-width: @{xs-min}) and (max-width: @{md-max})';
+@xs-lg: ~'(min-width: @{xs-min}) and (max-width: @{lg-max})';
+@sm: ~'(min-width: @{sm-min}) and (max-width: @{sm-max})';
+@sm-md: ~'(min-width: @{sm-min}) and (max-width: @{md-max})';
+@sm-lg: ~'(min-width: @{sm-min}) and (max-width: @{lg-max})';
+@sm-xl: ~'(min-width: @{sm-min}) and (max-width: @{xl-max})';
+@md: ~'(min-width: @{md-min}) and (max-width: @{md-max})';
+@md-lg: ~'(min-width: @{md-min}) and (max-width: @{lg-max})';
+@md-xl: ~'(min-width: @{md-min}) and (max-width: @{xl-max})';
+@lg: ~'(min-width: @{lg-min}) and (max-width: @{lg-max})';
+@lg-xl: ~'(min-width: @{lg-min}) and (max-width: @{xl-max})';
+@xl: ~'(min-width: @{xl-min}) and (max-width: @{xl-max})';
