+++ pages/src/common/colors.less
@@ -1,0 +1,74 @@
+//@ex-1: #0082ca;
+//@ex-2: #40bfef;
+
+/* Colors */
+@primary-blue: #0082ca;
+@primary-dark-blue: #0066ff;
+@primary-light-blue: #8bc2ff;
+
+@primary-orange: #ff5d24;
+@primary-orange-dark: #c6361a;
+
+@body-bg: #f8f8f8;
+@body-fg: rgba(22, 0, 27, 0.9);
+
+@link-fg: #0097E7;
+@link-hover-fg: #007fe7;
+@link-hover-underline: #ff5d24;
+
+@landing-bg-color: #131A2E;
+@landing-fg-color: #fff;
+
+@nav-dark-bg: @landing-bg-color; //#232429;
+@nav-dark-fg: #fff;
+@nav-dark-hover: #74c5ff;
+@nav-dark-selected: #00a2ff;
+@nav-dark-accent: #3eb2ff;
+
+// Color Helpers
+// Text helpers
+.text {
+  &-white {
+    color: #fff !important;
+  }
+  &-warning {
+    color: #ffc107 !important;
+  }
+  &-muted {
+    color: #6c757d !important;
+  }
+}
+
+// Backgrounds
+.bg {
+  &-info {
+    background-color: #17a2b8 !important;
+  }
+
+  &-danger {
+    background-color: #dc3545 !important;
+  }
+
+  &-white {
+    background-color: #fff !important;
+  }
+}
+
+// Borders
+.border {
+  border: 1px solid #dee2e6 !important;
+}
+
+.border-primary {
+  border-color: #007bff !important;
+}
+
+// Svg
+.fill {
+  &-light svg {
+    fill: #fff;
+  }
+  &-primary svg {
+    fill: @primary-blue;
+  }
+}
