--- pages/src/common/page.less
@@ -1,161 +1,0 @@
-html {
-  font-size: 16px;
-  font-family: sans-serif;
-  line-height: 1.15;
-  -webkit-text-size-adjust: 100%;
-  -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
-}
-
-body {
-  display: flex;
-  min-height: 100vh;
-  color: rgba(22, 0, 27, 0.9);
-  background: #f8f8f8;
-  flex-direction: column;
-}
-
-.brand-section {
-  width: 200px;
-}
-
-.error-details {
-  overflow: hidden auto;
-  white-space: normal;
-}
-
-main {
-  position: relative;
-  flex: 1 0 auto;
-}
-
-.page-scrollbar {
-  z-index: 100;
-}
-
-section {
-  margin-bottom: 20px;
-}
-
-.bg {
-  &-info {
-    background-color: #17a2b8 !important;
-  }
-
-  &-danger {
-    background-color: #dc3545 !important;
-  }
-
-  &-dark {
-    background-color: #343a40 !important;
-  }
-
-  &-white {
-    background-color: #fff !important;
-  }
-}
-
-button.bg-danger:hover,
-button.bg-danger:focus {
-  background-color: #bd2130 !important;
-}
-
-.border {
-  border: 1px solid #dee2e6 !important;
-}
-
-.border-primary {
-  border-color: #007bff !important;
-}
-
-.sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border: 0;
-}
-
-test-media-source,
-test-media-controls {
-  display: block;
-  padding: 5px;
-  text-align: center;
-  min-width: 250px;
-
-  .btn {
-    margin-right: 5px;
-  }
-}
-
-.demo-marquee {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-
-  position: relative;
-  padding: 40px;
-  min-height: 340px;
-
-  .bg-content {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-
-    z-index: 1;
-  }
-  .demo-content {
-    z-index: 2;
-
-    &.contrast {
-      color: white;
-      text-shadow: 0 0 4px rgba(0,0,0,.8);
-    }
-  }
-}
-
-.logo-light svg {
-  fill: #ffffff;
-}
-
-.logo-primary svg {
-  fill: #0082CA;
-}
-
-.unstyled-list li {
-  list-style-type: none;
-}
-
-.list-inline {
-  padding-left: 0;
-  list-style: none;
-
-  &-item {
-    display: inline-block;
-
-    &:not(:last-child) {
-      margin-right: 0.5rem;
-    }
-  }
-}
-
-.text {
-  &-center {
-    text-align: center !important;
-  }
-
-  &-white {
-    color: #fff !important;
-  }
-  &-warning {
-    color: #ffc107 !important;
-  }
-  &-muted {
-    color: #6c757d !important;
-  }
-}
