+++ pages/src/common/reboot.less
@@ -46,7 +46,9 @@ hr {
   border-top: 1px solid rgba(0, 0, 0, 0.1);
 }
 
-
+section {
+  margin-bottom: 20px;
+}
 
 pre,
 code {
