+++ pages/src/common/typography.less
@@ -1,5 +1,17 @@
 @import "./variables.less";
 
+html {
+  font-family: sans-serif;
+  line-height: 1.15;
+  -webkit-text-size-adjust: 100%;
+  -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+}
+
+body {
+  color: @body-fg;
+  background: @body-bg;
+}
+
 p {
   margin-top: 0;
   margin-bottom: 1rem;
@@ -14,20 +26,20 @@ h1, h2, h3, h4, h5,
 }
 
 h1, .h1 {
-  font-size: 3.5rem;
+  font-size: 2.5rem;
   @media @sm-xl {
-    font-size: 4.5rem;
+    font-size: 3rem;
     margin-bottom: 1rem;
   }
 }
 
 h2, .h2 {
-  font-size: 3rem;
+  font-size: 2rem;
   margin-bottom: 2rem;
 }
 
 h3, .h3 {
-  font-size: 2rem;
+  font-size: 1.75rem;
 }
 
 h4, .h4 {
