+++ pages/src/esl-media-demo/test-media.less
@@ -1,0 +1,39 @@
+test-media-source,
+test-media-controls {
+  display: block;
+  padding: 5px;
+  text-align: center;
+  min-width: 250px;
+
+  .btn {
+    margin-right: 5px;
+  }
+}
+
+.demo-marquee {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+
+  position: relative;
+  padding: 40px;
+  min-height: 340px;
+
+  .bg-content {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+
+    z-index: 1;
+  }
+  .demo-content {
+    z-index: 2;
+
+    &.contrast {
+      color: white;
+      text-shadow: 0 0 4px rgba(0,0,0,.8);
+    }
+  }
+}
