+++ pages/src/localdev.less
@@ -1,6 +1,6 @@
 @import "./common/reboot.less";
 @import "./common/typography.less";
-@import "./common/page.less";
+@import "./common/colors.less";
 @import "./common/markdown.less";
 
 @import "./common/accent-lines.less";
@@ -26,6 +26,7 @@
 @import "./benefits/benefits.less";
 @import "./browser-support/browser-support.less";
 @import "./overview/overview.less";
+@import "./esl-media-demo/test-media.less";
 
 @import "../../src/modules/all.less";
 @import "../../src/modules/draft/all.less";
