+++ pages/src/page/base-layout.less
@@ -1,0 +1,12 @@
+body {
+  display: flex;
+  min-height: 100vh;
+  flex-direction: column;
+}
+main {
+  position: relative;
+  flex: 1 0 auto;
+}
+.page-scrollbar {
+  z-index: 100;
+}
