+++ pages/src/page/landing-layout.less
@@ -1,4 +1,5 @@
 @import (reference) "../common/variables.less";
+@import "./base-layout.less";
 
 .landing-layout {
   height: 100vh;
