+++ pages/src/page/sidebar-layout.less
@@ -1,3 +1,6 @@
+@import (reference) "../common/variables.less";
+@import "./base-layout.less";
+
 .sidebar-layout {
   height: 100vh;
   display: flex;
