+++ pages/views/_includes/landing/marquee.njk
@@ -8,7 +8,7 @@
   </div>
 
   <div class="marquee-illustration" role="presentation">
-    <div class="marquee-brand logo-primary logo-animated">
+    <div class="marquee-brand logo-animated">
       {% include 'static/assets/logo-pretty.svg' %}
     </div>
   </div>
