+++ pages/views/_includes/navigation/header.njk
@@ -1,6 +1,6 @@
 <header class="header">
   <nav class="header-container container">
-    <a class="header-link header-logo logo-light logo-animated" href="{{ '/' | url }}" aria-label="Landing page">
+    <a class="header-link header-logo fill-light logo-animated" href="{{ '/' | url }}" aria-label="Landing page">
       {% include 'static/assets/logo.svg' %}
       <span class="header-short-name">esl</span>
       <span class="header-long-name">{{ site.name }}</span>
