+++ pages/views/_includes/navigation/sidebar.njk
@@ -3,7 +3,7 @@
 <esl-d-sidebar id="sidebar" class="open" aria-label="Site Navigation">
   <nav class="sidebar-nav">
     <a class="sidebar-heading" href="{{ '/' | url }}">
-      <span class="sidebar-logo logo-light logo-animated">{% include 'static/assets/logo.svg' %}</span>
+      <span class="sidebar-logo fill-light logo-animated">{% include 'static/assets/logo.svg' %}</span>
       <span class="sidebar-title">{{ site.name }}</span>
     </a>
