+++ pages/views/components/esl-scrollbar.njk
@@ -11,12 +11,3 @@ aside:
 ---
 
 {% mdRender 'src/modules/esl-scrollbar/README.md', 'intro' %}
-
-<section class="row">
-  <div class="col-12">
-    <h2>Related Examples: </h2>
-    <ul class="list-inline">
-      <li class="list-inline-item"><a href="{{ '/examples/scrollbar' | url }}">Scrollbar</a></li>
-    </ul>
-  </div>
-</section>
