+++ pages/views/utils/esl-a11ty-group.njk
@@ -11,12 +11,3 @@ aside:
 ---
 
 {% mdRender 'src/modules/esl-a11y-group/README.md', 'intro' %}
-
-<section class="row">
-  <div class="col-12">
-    <h2>Related Examples: </h2>
-    <ul class="list-inline">
-      <li class="list-inline-item"><a href="{{ '/examples/tabs' | url }}">Tabs</a></li>
-    </ul>
-  </div>
-</section>
